+++ src/main/java/com/stoqing/reservas/repository/OperarioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OperarioRepository extends JpaRepository<Operario, Long> {
+public interface OperarioRepository extends JpaRepository<Operario, Integer> {
     Optional<Operario> findByDni(String dni);
 }
